Implemented mechanism for executing and evaluating functions
The new feature can execute functions based on input tests and accordingly gives point to user
Also the EloRating system is modified, adapting it to the kind of problems where the user is dealing with some test cases.
For example: If a user completes 9/10, it's unfair to give all the points to the lesson, so we'll mark the user as losing to the lesson, but giving the lesson very few points, and subtracting from the user's elo rating, very few.